Feedback changes - restructed project structure
