+++ page-object/src/main/resources/sample-ui/album-list.html
