+++ page-object/src/main/resources/sample-ui/album-page.html
