+++ page-object/src/main/resources/sample-ui/css/album-list.css
