+++ page-object/src/main/resources/sample-ui/css/style.css
